HACK/requirements.txt
pymongo==4.4.0
python-dotenv==1.0.0
landingai-ade==0.2.0
pydantic==2.4.2
